fixed test cases

+++ lib/model/cart/pos_cart.dart
@@ -40,7 +40,7 @@ class PosCart {
     return totalMRP - totalFinalPrice;
   }
 
-  double get totalItemsCount {
+  int get totalItemsCount {
     return items.fold(0, (sum, item) => sum + item.quantity);
   }
 }
